backend/requirements.txt
django
djangorestframework
motor
pymongo
django-cors-headers
django-oauth-toolkit
python-dotenv
pytz
django-auth0